requirements.txt
python-dotenv
ccxt>=1.90.0
requests
fastapi
uvicorn
jinja2
python-multipart
gunicorn
pandas
pandas-ta
scikit-learn
joblib
websockets
scipy
xgboost
